Add files via upload
Results are now displayed in order from closest to farthest text neighbor, the index used for both searches is built using image vectors, and the number of neighbors for the text search is about 10x that of the image search. Overall, this seems to be working much better than the previous version.